+++ moment.d.ts
@@ -1,0 +1,2 @@
+import { MomentLocaleUtils } from './types/moment'
+export default MomentLocaleUtils