+++ src/DayPicker.js
@@ -15,7 +15,7 @@ import classNames from './classNames';
 import { ENTER, SPACE, LEFT, UP, DOWN, RIGHT } from './keys';
 
 export class DayPicker extends Component {
-  static VERSION = '7.2.4';
+  static VERSION = '7.3.0';
 
   static propTypes = {
     // Rendering months
@@ -381,16 +381,28 @@ export class DayPicker extends Component {
 
     switch (e.keyCode) {
       case LEFT:
-        this.showPreviousMonth();
+        if (this.props.dir === 'rtl') {
+          this.showNextMonth();
+        } else {
+          this.showPreviousMonth();
+        }
+        Helpers.cancelEvent(e);
         break;
       case RIGHT:
-        this.showNextMonth();
+        if (this.props.dir === 'rtl') {
+          this.showPreviousMonth();
+        } else {
+          this.showNextMonth();
+        }
+        Helpers.cancelEvent(e);
         break;
       case UP:
         this.showPreviousYear();
+        Helpers.cancelEvent(e);
         break;
       case DOWN:
         this.showNextYear();
+        Helpers.cancelEvent(e);
         break;
       default:
         break;
@@ -403,14 +415,23 @@ export class DayPicker extends Component {
 
   handleDayKeyDown = (day, modifiers, e) => {
     e.persist();
+
     switch (e.keyCode) {
       case LEFT:
         Helpers.cancelEvent(e);
-        this.focusPreviousDay(e.target);
+        if (this.props.dir === 'rtl') {
+          this.focusNextDay(e.target);
+        } else {
+          this.focusPreviousDay(e.target);
+        }
         break;
       case RIGHT:
         Helpers.cancelEvent(e);
-        this.focusNextDay(e.target);
+        if (this.props.dir === 'rtl') {
+          this.focusPreviousDay(e.target);
+        } else {
+          this.focusNextDay(e.target);
+        }
         break;
       case UP:
         Helpers.cancelEvent(e);