+++ types/moment.d.ts
@@ -1,0 +1,5 @@
+// TypeScript Version: 2.2
+
+import { LocaleUtils } from './utils';
+
+export const MomentLocaleUtils: LocaleUtils;